requirements.txt
hl7apy==1.3.4
aio-pika==9.0.7
fastapi==0.95.2
pydantic==1.10.8
python-dotenv==1.0.0
uvicorn==0.22.0
sqlalchemy==1.4.48
psycopg2-binary==2.9.6
PyJWT==2.7.0
passlib[bcrypt]